+++ federation/tests/test_fetchers.py
@@ -1,16 +1,23 @@
-# -*- coding: utf-8 -*-
 from unittest.mock import patch, Mock
 
-from federation.fetchers import retrieve_remote_profile
+from federation.fetchers import retrieve_remote_profile, retrieve_remote_content
 
 
-class TestRetrieveRemoteProfile(object):
+class TestRetrieveRemoteContent:
+    @patch("federation.fetchers.importlib.import_module")
+    def test_calls_diaspora_retrieve_and_parse_content(self, mock_import):
+        mock_retrieve = Mock()
+        mock_import.return_value = mock_retrieve
+        retrieve_remote_content("diaspora://user@example.com/status_message/1234", sender_key_fetcher=sum)
+        mock_retrieve.retrieve_and_parse_content.assert_called_once_with(
+            "diaspora://user@example.com/status_message/1234", sender_key_fetcher=sum,
+        )
+
+
+class TestRetrieveRemoteProfile:
     @patch("federation.fetchers.importlib.import_module")
     def test_calls_diaspora_retrieve_and_parse_profile(self, mock_import):
-        class MockRetrieve(Mock):
-            def retrieve_and_parse_profile(self, handle):
-                return "called with %s" % handle
-
-        mock_retrieve = MockRetrieve()
+        mock_retrieve = Mock()
         mock_import.return_value = mock_retrieve
-        assert retrieve_remote_profile("foo@bar") == "called with foo@bar"
+        retrieve_remote_profile("foo@bar")
+        mock_retrieve.retrieve_and_parse_profile.assert_called_once_with("foo@bar")